Move GlobablFlushingBehavior to seperate file
