+++ src/Voron/StorageEnvironment.cs
@@ -1,7 +1,6 @@
 ﻿using Sparrow;
 using Sparrow.Collections;
 using System;
-using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Diagnostics;
@@ -23,7 +22,6 @@ using Voron.Impl.FreeSpace;
 using Voron.Impl.Journal;
 using Voron.Impl.Paging;
 using Voron.Impl.Scratch;
-using Voron.Platform.Posix;
 using Voron.Platform.Win32;
 using Voron.Util;
 using Voron.Util.Conversion;
@@ -75,7 +73,7 @@ namespace Voron
         private readonly ConcurrentSet<LowLevelTransaction> _activeTransactions = new ConcurrentSet<LowLevelTransaction>();
 
         private readonly AbstractPager _dataPager;
-        private ExceptionDispatchInfo _flushingTaskFailure;
+        internal ExceptionDispatchInfo FlushingTaskFailure;
         private readonly WriteAheadJournal _journal;
         private readonly object _txWriter = new object();
         internal readonly ReaderWriterLockSlim FlushInProgressLock = new ReaderWriterLockSlim(LockRecursionPolicy.SupportsRecursion);
@@ -88,7 +86,7 @@ namespace Voron
         private CancellationTokenSource _cancellationTokenSource = new CancellationTokenSource();
         private readonly ScratchBufferPool _scratchBufferPool;
         private EndOfDiskSpaceEvent _endOfDiskSpace;
-        private int _sizeOfUnflushedTransactionsInJournalFile;
+        internal int SizeOfUnflushedTransactionsInJournalFile;
 
         private readonly Queue<TemporaryPage> _tempPagesPool = new Queue<TemporaryPage>();
         public bool Disposed;
@@ -403,7 +401,7 @@ namespace Voron
                     {
                         if (_endOfDiskSpace.CanContinueWriting)
                         {
-                            _flushingTaskFailure = null;
+                            FlushingTaskFailure = null;
                             _endOfDiskSpace = null;
                             _cancellationTokenSource = new CancellationTokenSource();
                             Task.Run(IdleFlushTimer);
@@ -493,7 +491,7 @@ namespace Voron
                 totalPages += page.NumberOfPages;
             }
 
-            Interlocked.Add(ref _sizeOfUnflushedTransactionsInJournalFile, totalPages);
+            Interlocked.Add(ref SizeOfUnflushedTransactionsInJournalFile, totalPages);
             if (tx.IsLazyTransaction == false)
                 GlobalFlusher.Value.MaybeFlushEnvironment(this);
         }
@@ -578,88 +576,7 @@ namespace Voron
             }
         }
 
-        private class GlobalFlushingBehavior
-        {
-            private readonly ConcurrentQueue<StorageEnvironment> _maybeNeedToFlush = new ConcurrentQueue<StorageEnvironment>();
-            private readonly ManualResetEventSlim _flushWriterEvent = new ManualResetEventSlim();
-            private readonly SemaphoreSlim _concurrentFlushes = new SemaphoreSlim(MaxConcurrentFlushes);
-
-            public void VoronEnvironmentFlushing()
-            {
-                // We want this to always run, even if we dispose / create new storage env, this is 
-                // static for the life time of the process, and environments will register / unregister from
-                // it as needed
-                while (true)
-                {
-                    _flushWriterEvent.Wait();
-                    _flushWriterEvent.Reset();
-
-                    StorageEnvironment envToFlush;
-                    while (_maybeNeedToFlush.TryDequeue(out envToFlush))
-                    {
-                        if (envToFlush.Disposed || envToFlush.Options.ManualFlushing)
-                            continue;
-
-                        var sizeOfUnflushedTransactionsInJournalFile = Volatile.Read(ref envToFlush._sizeOfUnflushedTransactionsInJournalFile);
-
-                        if (sizeOfUnflushedTransactionsInJournalFile == 0)
-                            continue; // nothing to do
-
-
-                        if (sizeOfUnflushedTransactionsInJournalFile <
-                            envToFlush._options.MaxNumberOfPagesInJournalBeforeFlush)
-                        {
-                            // we haven't reached the point where we have to flush, but we might want to, if we have enough 
-                            // resources available, if we have more than half the flushing capacity, we can do it now, otherwise, we'll wait
-                            // until it is actually required.
-                            if (_concurrentFlushes.CurrentCount < MaxConcurrentFlushes / 2)
-                                continue;
-                        }
-
-                        Interlocked.Add(ref envToFlush._sizeOfUnflushedTransactionsInJournalFile, -sizeOfUnflushedTransactionsInJournalFile);
-
-                        _concurrentFlushes.Wait();
-
-                        if (ThreadPool.QueueUserWorkItem(env =>
-                        {
-                            var storageEnvironment = ((StorageEnvironment)env);
-                            try
-                            {
-                                if (storageEnvironment.Disposed)
-                                    return;
-                                storageEnvironment.BackgroundFlushWritesToDataFile();
-                            }
-                            catch (Exception e)
-                            {
-                                storageEnvironment._flushingTaskFailure = ExceptionDispatchInfo.Capture(e.InnerException);
-                            }
-                            finally
-                            {
-                                _concurrentFlushes.Release();
-                            }
-                        }, envToFlush) == false)
-                        {
-                            _concurrentFlushes.Release();
-                            MaybeFlushEnvironment(envToFlush);// re-register if the thread pool is full
-                            Thread.Sleep(0); // but let it give up the execution slice so we'll let the TP time to run
-                        }
-                    }
-                }
-
-            }
-
-
-            public void MaybeFlushEnvironment(StorageEnvironment env)
-            {
-                _maybeNeedToFlush.Enqueue(env);
-                _flushWriterEvent.Set();
-            }
-        }
-
-
-
-
-        private void BackgroundFlushWritesToDataFile()
+        internal void BackgroundFlushWritesToDataFile()
         {
             try
             {
@@ -701,7 +618,7 @@ namespace Voron
 
         internal void AssertFlushingNotFailed()
         {
-            _flushingTaskFailure?.Throw(); // force re-throw of error
+            FlushingTaskFailure?.Throw(); // force re-throw of error
         }
 
         internal void HandleDataDiskFullException(DiskFullException exception)
